+++ src/contextMenu.item.directive.ts
@@ -6,8 +6,8 @@ import {Directive, Input, Output, EventEmitter, TemplateRef} from '@angular/core
   /* tslint:enable:directive-selector-type */
 })
 export class ContextMenuItemDirective {
-  @Input() public enabled: boolean = true;
-  @Input() public visible: boolean = true;
+  @Input() public enabled: boolean | ((item: any) => boolean) = true;
+  @Input() public visible: boolean | ((item: any) => boolean) = true;
   @Output() public execute: EventEmitter<{ event: Event, item: any }> = new EventEmitter<{ event: Event, item: any }>();
 
   constructor(private template: TemplateRef<{item: any}>) {}