+++ src/contextMenu.service.ts
@@ -1,8 +1,10 @@
-import {Injectable} from '@angular/core';
-import {Subject} from 'rxjs/Rx';
+import { ContextMenuComponent } from '../angular2-contextmenu';
+import { Injectable } from '@angular/core';
+import { Subject } from 'rxjs/Rx';
 
 export interface IContextMenuClickEvent {
   actions?: any[];
+  contextMenu?: ContextMenuComponent;
   event: MouseEvent;
   item: any;
 }